% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSMI.R
\name{calculateSMI}
\alias{calculateSMI}
\title{calculateSMI}
\usage{
calculateSMI(data, datatype = c("bulk", "sc"), species = c("human", "mouse"))
}
\arguments{
\item{data}{gene expression matrix (with rows as genes) or singlec-cell: seurat/sce objects}

\item{datatype}{is the data derived from bulk tissue or single-cell (sc)?}

\item{species}{between 'human' and 'mouse', select one based on whether the input expression data is derived from human or mouse}
}
\value{
a dataframe with rows as samples/cells, MYC Targets, PRC Targets (ssGSEA scores), SMI_unscaled, and SMI.
}
\description{
This function calculates the 'Stem Maturation Index' (SMI), a transcriptomic measure that indicates how well the sample or cell aligns along the stem-differentiation scale.
It also provides the MYC targets and PRC targets single-sample or single-cell scores.
}
\author{
Sudhir Malla
}
