% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highMeanGene.R
\name{highMeanGene}
\alias{highMeanGene}
\title{highMeanGene}
\usage{
highMeanGene(df)
}
\arguments{
\item{df}{dataframe with first column as genes}
}
\value{
expression matrix with unique genes as rows
}
\description{
Checks for duplicated genes from input expression data, selects gene with highest mean value from
duplicated genes, and produces expression matrix with uniques genes as rows.
}
