% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDSclassifier-package.R
\docType{package}
\name{PDSclassifier-package}
\alias{PDSclassifier}
\alias{PDSclassifier-package}
\title{PDSclassifier: Pathway-Derived Subtypes of Colorectal Cancer}
\description{
An pathway-based classifier of colorectal cancer that utilises pathway-level scores to stratify transcriptomic data into molecularly and biologically distinct subtypes, called Pathway-Derived Subtype (PDS). It can be applied to both human and mice.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dunne-lab.com/}
  \item \url{https://github.com/sidmall/PDSclassifier}
  \item Report bugs at \url{https://github.com/sidmall/PDSclassifier/issues}
}

}
\author{
\strong{Maintainer}: Sudhir Malla \email{s.malla@qub.ac.uk} (\href{https://orcid.org/0000-0001-5862-3815}{ORCID})

Authors:
\itemize{
  \item Philip Dunne \email{p.dunne@qub.ac.uk} (\href{https://orcid.org/0000-0001-9160-283X}{ORCID})
}

}
\keyword{internal}
