% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDSpredict.R
\name{PDSpredict}
\alias{PDSpredict}
\title{PDSpredict}
\usage{
PDSpredict(x, species = c("human", "mouse"), threshold = 0.6)
}
\arguments{
\item{x}{gene expression data in dataframe format with first column as genes}

\item{species}{between 'human' and 'mouse', select one based on whether the input expression data is derived from human or mouse}

\item{threshold}{set threshold between 0 to 1 for PDS calls, default 0.6. Those do not meet the threshold are indicated a 'mixed' bioloical subtype.}
}
\value{
a dataframe with Sample ID, PDS prediction scores for each subtye, PDS prediction and PDS prediction with threshold set by user (default 0.6)
}
\description{
PDSpredict gives PDS (Pathway Derived Subtypes) calls and prediction scores from input expression data
}
\author{
Sudhir Malla
}
