\name{testData}
\alias{testData}
\docType{data}
\title{
 testData
}
\description{
The dataset consists of gene expression profiles from 40 primary tumour CRC samples, with first column being genes.
}
\usage{data("testData")}
\format{
  A data frame with 27054 observations on the following 41 variables.
  \describe{
    \item{\code{Gene.Symbol}}{a character vector}
    \item{\code{SC00125A}}{a numeric vector}
    \item{\code{SC00131A}}{a numeric vector}
    \item{\code{SC00140A}}{a numeric vector}
    \item{\code{SC00146A}}{a numeric vector}
    \item{\code{SC00156A}}{a numeric vector}
    \item{\code{SC00159A}}{a numeric vector}
    \item{\code{SC00177A}}{a numeric vector}
    \item{\code{SC00182A}}{a numeric vector}
    \item{\code{SC00199A}}{a numeric vector}
    \item{\code{SC00208A}}{a numeric vector}
    \item{\code{SC00219A}}{a numeric vector}
    \item{\code{SC00259A}}{a numeric vector}
    \item{\code{SC00265A}}{a numeric vector}
    \item{\code{SC00268A}}{a numeric vector}
    \item{\code{SC00273A}}{a numeric vector}
    \item{\code{SC00280A}}{a numeric vector}
    \item{\code{SC00282A}}{a numeric vector}
    \item{\code{SC00290A}}{a numeric vector}
    \item{\code{SC00311A}}{a numeric vector}
    \item{\code{SC00314A}}{a numeric vector}
    \item{\code{SC00319A}}{a numeric vector}
    \item{\code{SC00326A}}{a numeric vector}
    \item{\code{SC00329A}}{a numeric vector}
    \item{\code{SC00334A}}{a numeric vector}
    \item{\code{SC00342A}}{a numeric vector}
    \item{\code{SC00346A}}{a numeric vector}
    \item{\code{SC00365A}}{a numeric vector}
    \item{\code{SC00400A}}{a numeric vector}
    \item{\code{SC00408A}}{a numeric vector}
    \item{\code{SC00410A}}{a numeric vector}
    \item{\code{SC00413A}}{a numeric vector}
    \item{\code{SC00419A}}{a numeric vector}
    \item{\code{SC00425A}}{a numeric vector}
    \item{\code{SC00429A}}{a numeric vector}
    \item{\code{SC00458A}}{a numeric vector}
    \item{\code{SC00497A}}{a numeric vector}
    \item{\code{SC00500A}}{a numeric vector}
    \item{\code{SC00507A}}{a numeric vector}
    \item{\code{SC00512A}}{a numeric vector}
    \item{\code{SC00521A}}{a numeric vector}
  }
}
\examples{
data(testData)
## maybe str(testData) ; plot(testData) ...
}
\keyword{datasets}
