% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDSpredict.R
\name{PDSpredict}
\alias{PDSpredict}
\title{PDSpredict}
\usage{
PDSpredict(x, species = c("human", "mouse"), threshold = 0.6)
}
\arguments{
\item{x}{gene expression data in dataframe format with first column as genes}

\item{species}{between 'human' and 'mouse', select one based on whether the input expression data is derived from human or mouse}

\item{threshold}{set threshold between 0 to 1 for PDS calls, default 0.6. Those do not meet the threshold are indicated a 'mixed' biological subtype.}
}
\value{
a dataframe with Sample ID, PDS prediction scores for each subtype, PDS prediction and PDS prediction with threshold set by user (default 0.6)
}
\description{
PDSpredict function provides a pathway-derived molecular classifications into three Pathway-Derived Subtypes (PDS): PDS1, PDS2, and PDS3.
It calculates PDS prediction probability scores as an indication of how heterogeneous a sample is with these PDS-defined biology.
The default threshold of 0.6 has been applied to get PDS calls from the prediction probability scores.
}
\author{
Sudhir Malla
}
